+++ changelog.md
@@ -481,6 +481,7 @@ are 9001 for HSQLDB, 9092 for H2 and 1527 for Derby.
 - Remove MRJToolkitStubs.jar
 - Add an audit mechanism to print performances (times) for assignments
 - Improved cost parser, which is now between 5 and 20 times faster
+- Improved multi-flow edge update check